requirements.txt
pdfplumber
transformers
sentence-transformers
langchain
streamlit
tqdm
python-dotenv
ibm-watson-machine-learning
